ADDED PUPPET MODULE MOTD
MOTD module greets user with a message of the day when logging in to the terminal. This helps the admin keep track of what system they are signing into. MOTD module also displays useful information about the OS and networking of the system to make it easier for the admin to see what system they are working with.